+++ examples/peg_solitaire/PegSolitaire.h
@@ -10,7 +10,7 @@ public:
     [[nodiscard]] size_t get_action_count() const override;
     [[nodiscard]] int get_score() const override;
     StepResult get_step(int index) override;
-    [[nodiscard]] std::vector<double> get_state() const override;
+    [[nodiscard]] std::vector<float> get_state() const override;
     void reset() override;
 
     typedef enum peg_status
