+++ mlgames/AgentQTrainer.cpp
@@ -3,7 +3,7 @@
 
 AgentQTrainer::AgentQTrainer(
         LinearNN *model_, torch::optim::Optimizer *optimizer_, const c10::DeviceType device_,
-        size_t batch_size_, double gamma_)
+        size_t batch_size_, float gamma_)
 {
     device = device_;
     model = model_;
@@ -14,9 +14,9 @@ AgentQTrainer::AgentQTrainer(
 
 
 std::vector<int>
-AgentQTrainer::get_play(const std::vector<double> &state, size_t count_samples) const
+AgentQTrainer::get_play(const std::vector<float> &state, size_t count_samples) const
 {
-    const auto state0 = torch::tensor(state, torch::kDouble)
+    const auto state0 = torch::tensor(state, torch::kFloat)
                                 .reshape({(long) count_samples, (long) (*model)->input_size})
                                 .to(device);
     (*model)->eval();
@@ -39,7 +39,7 @@ AgentQTrainer::get_play(const std::vector<double> &state, size_t count_samples) 
 
 
 std::vector<int>
-AgentQTrainer::get_action(const std::vector<double> &state, size_t count_samples) const
+AgentQTrainer::get_action(const std::vector<float> &state, size_t count_samples) const
 {
     const int epsilon = 80 - number_of_games;
 
@@ -57,16 +57,16 @@ AgentQTrainer::get_action(const std::vector<double> &state, size_t count_samples
 
 
 void AgentQTrainer::train_short_memory(
-        const std::vector<double> &state, const std::vector<int> &action, int reward,
-        const std::vector<double> &next_state, bool game_over)
+        const std::vector<float> &state, const std::vector<int> &action, int reward,
+        const std::vector<float> &next_state, bool game_over)
 {
     train_step(1, state, action, {reward}, next_state, {game_over});
 }
 
 
 void AgentQTrainer::remember(
-        const std::vector<double> &state, const std::vector<int> &action, const int reward,
-        const std::vector<double> &next_state, const bool game_over)
+        const std::vector<float> &state, const std::vector<int> &action, const int reward,
+        const std::vector<float> &next_state, const bool game_over)
 {
     memory_deque.push_back(
             {(*model)->input_size, (*model)->output_size, state, action, reward, next_state,
@@ -98,10 +98,10 @@ void AgentQTrainer::train_long_memory()
     const size_t num_state = samples[0].num_state;
     const size_t num_action = samples[0].num_action;
 
-    std::vector<double> states;
+    std::vector<float> states;
     std::vector<int> actions;
     std::vector<int> rewards;
-    std::vector<double> next_states;
+    std::vector<float> next_states;
     std::vector<bool> game_overs;
 
     states.reserve(sample_size * num_state);
@@ -124,20 +124,20 @@ void AgentQTrainer::train_long_memory()
 }
 
 void AgentQTrainer::train_step(
-        size_t count_samples, const std::vector<double> &old_states_,
+        size_t count_samples, const std::vector<float> &old_states_,
         const std::vector<int> &actions_, const std::vector<int> &rewards_,
-        const std::vector<double> &new_states_, const std::vector<bool> &dones_)
+        const std::vector<float> &new_states_, const std::vector<bool> &dones_)
 {
     const torch::Tensor old_states =
-            torch::tensor(old_states_, torch::kDouble)
+            torch::tensor(old_states_, torch::kFloat)
                     .reshape({(long) count_samples, (long) (*model)->input_size})
                     .to(device);
     const torch::Tensor actions =
             torch::tensor(actions_, torch::kInt)
                     .reshape({(long) count_samples, (long) (*model)->output_size})
-                    .to(device, torch::kDouble);
+                    .to(device, torch::kFloat);
     const torch::Tensor new_states =
-            torch::tensor(new_states_, torch::kDouble)
+            torch::tensor(new_states_, torch::kFloat)
                     .reshape({(long) count_samples, (long) (*model)->input_size})
                     .to(device);
 
@@ -157,7 +157,7 @@ void AgentQTrainer::train_step(
             if (!dones_[index])
             {
                 q_new = q_new +
-                        gamma * (*model)->forward(new_states[index])[max_index].item().toDouble();
+                        gamma * (*model)->forward(new_states[index])[max_index].item().toFloat();
             }
             target[index][max_index] = q_new;
         }
