+++ mlgames/include/mlgames/AbstractGame.h
@@ -18,6 +18,6 @@ public:
     [[nodiscard]] virtual size_t get_action_count() const = 0;
     [[nodiscard]] virtual int get_score() const = 0;
     virtual StepResult get_step(int action) = 0;
-    [[nodiscard]] virtual std::vector<double> get_state() const = 0;
+    [[nodiscard]] virtual std::vector<float> get_state() const = 0;
     virtual void reset() = 0;
 };
